refactor: Split auth & documents
